% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{DelMissIVF}
\alias{DelMissIVF}
\title{Delete variables with missing values for IVF-ET data}
\usage{
DelMissIVF(PID, OutPath = "default")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}
}
\value{
An R6 class object containing the variable(s) without missing values.
}
\description{
Delete missing variables with low variance
}
\examples{
res <- InitIVF()
    res1 = LoadIVF(PID = res$PID, UseExample = "example#1")
    res2 = DelMissIVF(PID=res$PID)
    FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
