% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{FuncExit}
\alias{FuncExit}
\title{End the module analysis}
\usage{
FuncExit(PID)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}
}
\description{
End the module analysis
}
\examples{
   res <- InitIVF()
   res1 = LoadIVF(PID = res$PID, UseExample = "example#1")
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
