% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{InitIVF}
\alias{InitIVF}
\title{Initialize IVFPred module}
\usage{
InitIVF()
}
\value{
An R6 class object.
}
\description{
Initialize IVFPred module analysis. It can generate an R6 class object.
}
\details{
IVFPred module is designed to improve model efficiency in
predicting pregnancy success rates in IVF-ET and screening
the key influencing factors for the individual centers and
patients.
}
\examples{
res <- InitIVF()
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
