% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{IVFPred}
\alias{IVFPred}
\title{Build predict model for IVF-ET data}
\usage{
IVFPred(PID, OutPath = "default", AutTuneM = "random_search", AutTuneN = "5", RsmpMethod = "cv", Folds = "5", Ratio = NULL, Repeats = NULL, Learners = NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{AutTuneM}{chr. Method for hyper-parameter autotuning.
Options include "default", "random_search", "grid_search", "nloptr"(Non-linear optimization), and "gensa"(Generalized simulated annealing).
The "default" option uses the simple training method for parameter optimization of mlr3 package.}

\item{AutTuneN}{num. Upper limit of model tuning times. It should be more than 20 times to search the appropriate parameters,
but it takes more time. In theory, more time, better training results.}

\item{RsmpMethod}{chr. Method for resampling. Options include "cv"(cross validation), "loo"(leave-one-out cross validation),
"bootstrap"(bootstrapping), "holdout"(holdout).}

\item{Folds}{num. Folds for cross validation resampling method. The default value is 5.}

\item{Ratio}{num. Ratio for "Holdout" resampling method. The default value is 5.}

\item{Repeats}{num. Repeats for "Bootstrap" resampling method.}

\item{Learners}{chr. Learners for build IVF-ET outcome predicting model. Options include "lasso", "elastic net", "rf"(Random forest), "stacked generalization", "logistic" and "xgboost"(Xgboost).
One or more arbitrary options can be selected at the same time. Note that separates different learners by "," and without space(e.g. Learners ="lasso,rf,xgboost").}
}
\value{
An R6 class object containing the model results
}
\description{
Build predict model for IVF-ET data
}
\examples{
    res <- InitIVF()
   res1 = LoadIVF(PID = res$PID, UseExample = "example#2")
   res2 = TransImputIVF(PID=res$PID, Vars="all.x")
   res2 = DelLowVarIVF(PID=res$PID)
   res2 = DelMissIVF(PID=res$PID)
   res2 = TransTypeIVF(PID = res$PID, Vars = "Y,X85,X86,X87,X88,X89,X90,X91,X92,X93,X94,X95,X96,X97,X98",
                       To = "factor")
   res2 = TransClassIVF(PID = res$PID, Vars = "X47", LevelTo = 4)
   res2 = TransScaleIVF(PID = res$PID, Vars = "X55", Method = "normal")
   res2 = TransDistrIVF(PID = res$PID, Vars = "X61", Method = "log10")
   res3 = IVFPred(PID = res$PID, AutTuneM = "random_search", AutTuneN = 5,
                  RsmpMethod = "cv", Folds = 5,
                  Learners = "logistic,rf,xgboost")
   FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
