% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{LoadIVF}
\alias{LoadIVF}
\title{Load data file for IVFPred module}
\usage{
LoadIVF(PID, UseExample = "default", DataPath = NULL, VocaPath = NULL)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by IVFPred}

\item{UseExample}{chr. Method of uploading data. If "default",user should upload their own data files,
or use "example#1" or "example#2" provided by this module.}

\item{DataPath}{chr. Input directory of data file, e.g. "D:/test/eg_data_IVFPred.xlsx". It should be
noted that the slash symbol is "/", not "\".}

\item{VocaPath}{chr. Input directory of vocabulary file, e.g. "D:/test/eg_voca_IVFPred.xlsx". It should be
noted that the slash symbol is "/", not "\".}
}
\value{
An R6 class object containing the input data.
}
\description{
Load data file for IVFPred module
}
\examples{
res <- InitIVF()
   res = LoadIVF(PID = res$PID, UseExample = "example#1")
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
