% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{TransDistrIVF}
\alias{TransDistrIVF}
\title{Transform variable distribution for IVF-ET data}
\usage{
TransDistrIVF(PID,OutPath = "default", Vars, Method)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables;
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{Method}{chr. Methods used for imputation. Available options include "lod" or "cart".
For "lod" method, limit of detection (LOD) should be included in the "Vocabulary" file.}
}
\value{
An R6 class object containing the variable(s) after transforming distribution.
}
\description{
Transform variable distribution
}
\examples{
res <- InitIVF()
    res1 = LoadIVF(PID = res$PID, UseExample = "example#1")
    res2 = TransDistrIVF(PID = res$PID, Vars = "X61", Method = "log10")
    FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
