% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{TransImputIVF}
\alias{TransImputIVF}
\title{Missing data imputation for IVF-ET data}
\usage{
TransImputIVF(PID,OutPath = "default",Vars)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables;
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}
}
\value{
An R6 class object containing variable(s) with imputation.
}
\description{
Missing data imputation.
}
\examples{
res <- InitIVF()
    res1 = LoadIVF(PID = res$PID, UseExample = "example#1")
    res2 = TransImputIVF(PID=res$PID, Vars="all.x")
    FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
