% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{TransScaleIVF}
\alias{TransScaleIVF}
\title{Scale variables for IVF-ET data}
\usage{
TransScaleIVF(PID,OutPath = "default", Vars, Method = "normal",Direct="positive",
      RangeLow="0", RangeUpper="1")
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables;
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{Method}{chr. Scaling methods. Available options include "normal" and "range".}

\item{Direct}{chr. Direction to be transformed, Available options include "positive" and "negative".}

\item{RangeLow}{num. Lower limit for range method.}

\item{RangeUpper}{num. Upper limit for range method. It should be greater than the lower limit.}
}
\value{
An R6 class object containing the variable(s) after scaling data.
}
\description{
Scale variables
}
\examples{
res <- InitIVF()
    res1 = LoadIVF(PID = res$PID, UseExample = "example#1")
    res2 = TransScaleIVF(PID = res$PID, Vars = "X55", Method = "normal")
    FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
