% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVFPred.R
\name{TransTypeIVF}
\alias{TransTypeIVF}
\title{Transform data type for IVF-ET data}
\usage{
TransTypeIVF(PID,OutPath = "default",Vars,To)
}
\arguments{
\item{PID}{chr. Program ID. It must be the same with the PID generated by initial functions.}

\item{OutPath}{chr. Output file directory, e.g. "D:/test". It should be noted that
the slash symbol is "/", not "\". If "default", the current working directory will be set.}

\item{Vars}{Variables to be imputed. Available options include:
"all.x", all exposure variables;
It should be noted that there is fixed format for the entering characters
separated with comma and without space, e.g., "X1,X2,X3".}

\item{To}{chr. Indicate the type of the chosen variables to be transformed into.
Available options include "integer", "numeric", "character", "factor", "logical", and "date".}
}
\value{
An R6 class object containing the variable(s) after transforming data type.
}
\description{
Transform data type
}
\examples{
res <- InitIVF()
    res1 = LoadIVF(PID = res$PID, UseExample = "example#1")
    res2 = TransTypeIVF(PID = res$PID, Vars = "all.x", To = "character")
    FuncExit(PID = res$PID)
}
\author{
Weinan Lin, Bin Wang (corresponding author)
}
